+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-## 1.0.0
+## 1.0.0+1
 
-- Initial version, created by Stagehand
+- Update RxDart package to ^0.21.0, DistinctValueConnectableObservable now has a separate factory constructor seeded().
+  This allows you to seed this stream with a null value.
 
 ## 1.0.0+1
 
 - Some refactors, update readme.md
+
+## 1.0.0
+
+- Initial version, created by Stagehand
+